refactor(tests): split command corpus oracle
