+++ tests/fixtures/guard-command-corpus/seed-manifest.json
@@ -7,7 +7,13 @@
   "fixed_seed": 22020,
   "permutation_multiplier": 7919,
   "generator_source_sha256": "6e2d16b8666a755b2f12d99b86774e7315cf2f349625dd472a22a728133919f1",
-  "oracle_source_sha256": "5a825c30932e119182aaf12c4d4f3b0c9e88ca415a36315880cf314f49dff26d",
+  "oracle_source_sha256": {
+    "guard_command_corpus_oracle.py": "7c7dee62735838dd5cf759156a64860e24c1afc6a5df54aeb402a9785f1a0d91",
+    "guard_command_corpus_oracle_types.py": "31ae3f338de752bf5d122e9a8274184bc525c1fc4dbf4410f970bb380628e59c",
+    "guard_command_corpus_oracle_benign.py": "a857e5b8d4c4abce7e92c7b7f66f6440ee269a0c5201c88ed23bd4a40eb6e5b6",
+    "guard_command_corpus_oracle_adversarial.py": "cbe6761e3ae7e6f646947d212998a26846742ccd702f2d6b237c72f83b8d23c5",
+    "guard_command_corpus_oracle_loader.py": "22e766e8fe437c675ac38256433b2a47dfe122ad06c5d3d60f88eaccdefeec48"
+  },
   "pairs_sha256": "73da046da2c4d9f0c45b6dc70e3db3e2045199c7aa5e58e2516086d490fe9c11",
   "known_gaps_sha256": "c728febfbc8852913826a3fbf11ed654c044c5412cf1b55d07e64b8b20e1c5a2",
   "reproducible_command": "uv run --no-sync pytest -q tests/test_guard_command_corpus.py",
