+++ tests/guard_command_corpus_oracle_adversarial.py
@@ -1,0 +1,287 @@
+"""Reviewed adversarial and minimal-pair oracle data."""
+
+from __future__ import annotations
+
+from tests.guard_command_corpus_oracle_benign import PROCESS_READ_EFFECTS, READ_EFFECTS
+from tests.guard_command_corpus_oracle_types import OracleSeed, PairOracleFacts
+
+ADVERSARIAL_ORACLE: dict[str, OracleSeed] = {
+    "flag-order": OracleSeed(
+        "network-container-cloud",
+        ("network-write", "remote-state-mutation", "destructive-or-irreversible-operation"),
+        "remote-resource",
+        (),
+        ("operation-and-targets", "remote-resource-identity"),
+        "block",
+        "decidable",
+        "CDX-064",
+    ),
+    "flag-values": OracleSeed(
+        "credentials-permissions-system-guard-destruction",
+        ("network-write", "remote-state-mutation", "destructive-or-irreversible-operation"),
+        "remote-resource",
+        (),
+        ("operation-and-targets", "remote-resource-identity"),
+        "block",
+        "decidable",
+        "CDX-064",
+    ),
+    "wrappers": OracleSeed(
+        "shell-composition",
+        ("process-execution", "workspace-write", "destructive-or-irreversible-operation"),
+        "workspace",
+        ("dynamic-input",),
+        ("launch-chain", "shell-data-flow", "workspace-identity"),
+        "block",
+        "uncertain",
+        "CDX-066",
+    ),
+    "redirects": OracleSeed(
+        "network-container-cloud",
+        ("network-write", "remote-state-mutation", "destructive-or-irreversible-operation"),
+        "remote-resource",
+        (),
+        ("operation-and-targets", "shell-data-flow", "remote-resource-identity"),
+        "block",
+        "decidable",
+        "CDX-064",
+    ),
+    "symlinks": OracleSeed(
+        "workspace-patch-write",
+        ("external-filesystem-write", "guard-control-operation"),
+        "guard",
+        ("unresolved-launch-identity",),
+        ("operation-and-targets", "workspace-identity"),
+        "block",
+        "uncertain",
+        "CDX-062",
+    ),
+    "path-shadowing": OracleSeed(
+        "package-runner-source",
+        ("process-execution", "guard-control-operation"),
+        "guard",
+        ("unresolved-launch-identity",),
+        ("launch-chain", "executable-identity"),
+        "require-reapproval",
+        "uncertain",
+        "CDX-065",
+    ),
+    "heredocs": OracleSeed(
+        "shell-composition",
+        ("process-execution", "workspace-write", "destructive-or-irreversible-operation"),
+        "workspace",
+        ("dynamic-input",),
+        ("shell-data-flow", "workspace-identity"),
+        "block",
+        "uncertain",
+        "CDX-066",
+    ),
+    "substitutions": OracleSeed(
+        "source-search-read",
+        ("process-execution", "sensitive-read"),
+        "sensitive-local",
+        ("dynamic-input",),
+        ("shell-data-flow", "operation-and-targets"),
+        "require-reapproval",
+        "uncertain",
+        "CDX-066",
+    ),
+    "pipelines": OracleSeed(
+        "shell-composition",
+        ("process-execution", "sensitive-read", "external-filesystem-write"),
+        "external-local",
+        ("dynamic-input",),
+        ("shell-data-flow", "operation-and-targets"),
+        "require-reapproval",
+        "uncertain",
+        "CDX-066",
+    ),
+    "package-sources": OracleSeed(
+        "package-runner-source",
+        ("process-execution", "package-or-source-installation", "workspace-write"),
+        "workspace",
+        ("unresolved-launch-identity",),
+        ("executable-identity", "dependency-provenance", "launch-chain"),
+        "require-reapproval",
+        "uncertain",
+        "CDX-065",
+    ),
+    "graphql": OracleSeed(
+        "github-remote",
+        ("network-write", "remote-state-mutation", "destructive-or-irreversible-operation"),
+        "remote-resource",
+        (),
+        ("operation-and-targets", "remote-resource-identity"),
+        "block",
+        "decidable",
+        "CDX-064",
+    ),
+    "capability-replay": OracleSeed(
+        "credentials-permissions-system-guard-destruction",
+        ("guard-control-operation",),
+        "guard",
+        ("dynamic-input",),
+        ("capability-constraints", "operation-and-targets"),
+        "block",
+        "uncertain",
+        "CDX-066",
+    ),
+}
+
+
+PAIR_ORACLE: dict[str, tuple[PairOracleFacts, PairOracleFacts]] = {
+    "p-navigation-boundary": (
+        PairOracleFacts(READ_EFFECTS, "workspace", "reversible", (), ("working-directory-identity",), "review"),
+        PairOracleFacts(
+            READ_EFFECTS,
+            "external-local",
+            "reversible",
+            (),
+            ("working-directory-identity", "operation-and-targets"),
+            "require-reapproval",
+        ),
+    ),
+    "p-source-boundary": (
+        PairOracleFacts(READ_EFFECTS, "workspace", "reversible", (), ("workspace-identity",), "review"),
+        PairOracleFacts(
+            READ_EFFECTS,
+            "external-local",
+            "reversible",
+            (),
+            ("workspace-identity", "operation-and-targets"),
+            "require-reapproval",
+        ),
+    ),
+    "p-typescript-source": (
+        PairOracleFacts(
+            PROCESS_READ_EFFECTS,
+            "workspace",
+            "reversible",
+            (),
+            ("executable-identity", "dependency-provenance"),
+            "review",
+        ),
+        PairOracleFacts(
+            ("process-execution", "package-or-source-installation", "workspace-write"),
+            "workspace",
+            "recoverable-with-review",
+            ("unresolved-launch-identity",),
+            ("executable-identity", "dependency-provenance", "launch-chain"),
+            "require-reapproval",
+        ),
+    ),
+    "p-git-history": (
+        PairOracleFacts(READ_EFFECTS, "workspace", "reversible", (), ("repository-identity",), "review"),
+        PairOracleFacts(
+            ("process-execution", "workspace-write", "destructive-or-irreversible-operation"),
+            "workspace",
+            "recoverable-with-review",
+            (),
+            ("repository-identity", "operation-and-targets"),
+            "block",
+        ),
+    ),
+    "p-github-mutation-impact": (
+        PairOracleFacts(
+            ("network-write", "remote-state-mutation"),
+            "remote-resource",
+            "recoverable-with-review",
+            (),
+            ("remote-resource-identity", "operation-and-targets"),
+            "review",
+        ),
+        PairOracleFacts(
+            ("network-write", "remote-state-mutation", "destructive-or-irreversible-operation"),
+            "remote-resource",
+            "irreversible",
+            (),
+            ("remote-resource-identity", "operation-and-targets"),
+            "block",
+        ),
+    ),
+    "p-package-operation": (
+        PairOracleFacts(
+            ("process-execution", "network-read", "remote-state-read"),
+            "remote-resource",
+            "reversible",
+            (),
+            ("executable-identity", "remote-resource-identity"),
+            "review",
+        ),
+        PairOracleFacts(
+            ("process-execution", "network-read", "package-or-source-installation", "workspace-write"),
+            "workspace",
+            "recoverable-with-review",
+            (),
+            ("executable-identity", "dependency-provenance", "workspace-identity"),
+            "require-reapproval",
+        ),
+    ),
+    "p-shell-data-vs-eval": (
+        PairOracleFacts(
+            ("process-execution", "workspace-or-public-read"),
+            "public-resource",
+            "reversible",
+            (),
+            ("operation-and-targets",),
+            "review",
+        ),
+        PairOracleFacts(
+            ("process-execution",),
+            "unknown",
+            "unknown",
+            ("dynamic-input",),
+            ("shell-data-flow", "expected-effects"),
+            "require-reapproval",
+        ),
+    ),
+    "p-cloud-help-redirection": (
+        PairOracleFacts(
+            ("process-execution", "workspace-or-public-read"),
+            "public-resource",
+            "reversible",
+            (),
+            ("operation-and-targets",),
+            "review",
+        ),
+        PairOracleFacts(
+            ("process-execution", "network-write", "remote-state-mutation", "destructive-or-irreversible-operation"),
+            "remote-resource",
+            "irreversible",
+            (),
+            ("shell-data-flow", "remote-resource-identity"),
+            "block",
+        ),
+    ),
+    "p-patch-check-vs-apply": (
+        PairOracleFacts(
+            PROCESS_READ_EFFECTS,
+            "workspace",
+            "reversible",
+            (),
+            ("workspace-identity", "operation-and-targets"),
+            "review",
+        ),
+        PairOracleFacts(
+            ("process-execution", "workspace-write"),
+            "workspace",
+            "reversible",
+            (),
+            ("workspace-identity", "operation-and-targets"),
+            "review",
+        ),
+    ),
+    "p-capability-replay": (
+        PairOracleFacts(
+            ("guard-control-operation",), "guard", "recoverable-with-review", (), ("capability-constraints",), "review"
+        ),
+        PairOracleFacts(
+            ("guard-control-operation",),
+            "guard",
+            "recoverable-with-review",
+            ("dynamic-input",),
+            ("capability-constraints",),
+            "block",
+        ),
+    ),
+}
