+++ tests/guard_command_corpus_oracle_benign.py
@@ -1,0 +1,428 @@
+"""Reviewed benign oracle seed data."""
+
+from __future__ import annotations
+
+from typing import cast
+
+from tests.guard_command_corpus_oracle_types import DecisionStatus, OracleFloor, OracleSeed
+
+READ_EFFECTS = ("process-execution", "workspace-or-public-read")
+PROCESS_READ_EFFECTS = ("process-execution", "workspace-or-public-read")
+
+# Every reviewed seed has an explicit entry. Broad workflow-name inference is forbidden.
+BENIGN_ORACLE: dict[str, OracleSeed] = {
+    **{
+        source: OracleSeed(
+            workflow,
+            effects,
+            target,
+            (),
+            proofs,
+            cast(OracleFloor, floor),
+            cast(DecisionStatus, status),
+            owner,
+        )
+        for source, workflow, effects, target, proofs, floor, status, owner in (
+            (
+                "working-directory",
+                "navigation-public-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "working-directory-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "directory-list",
+                "navigation-public-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "repository-root",
+                "navigation-public-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "repository-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "status",
+                "navigation-public-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "repository-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "source-files",
+                "source-search-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "targeted-search",
+                "source-search-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "source-view",
+                "source-search-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "symbol-search",
+                "source-search-read",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "typescript-regression",
+                "build-typecheck-test-lint",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("executable-identity", "dependency-provenance", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "pytest",
+                "build-typecheck-test-lint",
+                ("process-execution", "workspace-or-public-read", "workspace-write"),
+                "workspace",
+                ("executable-identity", "workspace-identity", "expected-effects"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "bun-build",
+                "build-typecheck-test-lint",
+                ("process-execution", "workspace-or-public-read", "workspace-write"),
+                "workspace",
+                ("executable-identity", "dependency-provenance", "workspace-identity", "expected-effects"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "ruff",
+                "build-typecheck-test-lint",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("executable-identity", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "diff-check",
+                "git-local",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "repository-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "recent-log",
+                "git-local",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "repository-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "show-stat",
+                "git-local",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "repository-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "branches",
+                "git-local",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "repository-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "resolve-thread-regression",
+                "github-remote",
+                ("network-write", "remote-state-mutation"),
+                "remote-resource",
+                ("operation-and-targets", "remote-resource-identity"),
+                "review",
+                "context-required",
+                "CDX-063",
+            ),
+            (
+                "merge-regression",
+                "github-remote",
+                ("network-write", "remote-state-mutation"),
+                "remote-resource",
+                ("operation-and-targets", "remote-resource-identity"),
+                "require-reapproval",
+                "decidable",
+                "CDX-064",
+            ),
+            (
+                "pr-view",
+                "github-remote",
+                ("network-read", "remote-state-read"),
+                "remote-resource",
+                ("operation-and-targets", "remote-resource-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "pr-checks",
+                "github-remote",
+                ("network-read", "remote-state-read"),
+                "remote-resource",
+                ("operation-and-targets", "remote-resource-identity"),
+                "review",
+                "context-required",
+                "CDX-060",
+            ),
+            (
+                "bun-lock",
+                "package-runner-source",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("executable-identity", "dependency-provenance", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "uv-tree",
+                "package-runner-source",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("executable-identity", "dependency-provenance", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "cargo-tree",
+                "package-runner-source",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("executable-identity", "dependency-provenance", "workspace-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "registry-source",
+                "package-runner-source",
+                ("process-execution", "network-read", "remote-state-read"),
+                "remote-resource",
+                ("executable-identity", "operation-and-targets", "remote-resource-identity"),
+                "review",
+                "context-required",
+                "CDX-065",
+            ),
+            (
+                "compose-ps",
+                "network-container-cloud",
+                ("process-execution", "system-or-privilege-operation", "remote-state-read"),
+                "system",
+                ("executable-identity", "configuration-identity", "operation-and-targets"),
+                "review",
+                "context-required",
+                "CDX-066",
+            ),
+            (
+                "container-inspect",
+                "network-container-cloud",
+                ("process-execution", "system-or-privilege-operation", "remote-state-read"),
+                "system",
+                ("executable-identity", "operation-and-targets"),
+                "review",
+                "context-required",
+                "CDX-066",
+            ),
+            (
+                "aws-identity",
+                "network-container-cloud",
+                ("process-execution", "network-read", "remote-state-read", "credential-or-secret-operation"),
+                "remote-resource",
+                ("executable-identity", "configuration-identity", "remote-resource-identity"),
+                "review",
+                "context-required",
+                "CDX-066",
+            ),
+            (
+                "gcloud-project",
+                "network-container-cloud",
+                ("process-execution", "network-read", "remote-state-read", "credential-or-secret-operation"),
+                "remote-resource",
+                ("executable-identity", "configuration-identity", "remote-resource-identity"),
+                "review",
+                "context-required",
+                "CDX-066",
+            ),
+            (
+                "patch-check",
+                "workspace-patch-write",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("executable-identity", "workspace-identity", "operation-and-targets"),
+                "review",
+                "context-required",
+                "CDX-062",
+            ),
+            (
+                "patch-apply",
+                "workspace-patch-write",
+                ("process-execution", "workspace-write"),
+                "workspace",
+                ("executable-identity", "workspace-identity", "operation-and-targets"),
+                "review",
+                "context-required",
+                "CDX-062",
+            ),
+            (
+                "format-write",
+                "workspace-patch-write",
+                ("process-execution", "workspace-write"),
+                "workspace",
+                ("executable-identity", "workspace-identity", "operation-and-targets"),
+                "review",
+                "context-required",
+                "CDX-062",
+            ),
+            (
+                "copy-generated",
+                "workspace-patch-write",
+                ("process-execution", "workspace-write"),
+                "workspace",
+                ("executable-identity", "workspace-identity", "operation-and-targets"),
+                "review",
+                "context-required",
+                "CDX-062",
+            ),
+            (
+                "cd-pipeline",
+                "shell-composition",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("shell-data-flow", "workspace-identity", "executable-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "stderr-pipeline",
+                "shell-composition",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("shell-data-flow", "workspace-identity", "executable-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "find-exec",
+                "shell-composition",
+                ("process-execution", "workspace-or-public-read", "workspace-write"),
+                "workspace",
+                ("shell-data-flow", "workspace-identity", "executable-identity", "expected-effects"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "json-pipeline",
+                "shell-composition",
+                PROCESS_READ_EFFECTS,
+                "workspace",
+                ("shell-data-flow", "repository-identity", "executable-identity"),
+                "review",
+                "context-required",
+                "CDX-061",
+            ),
+            (
+                "credential-metadata",
+                "credentials-permissions-system-guard-destruction",
+                ("credential-or-secret-operation", "sensitive-read"),
+                "sensitive-local",
+                ("operation-and-targets", "executable-identity"),
+                "require-reapproval",
+                "decidable",
+                "CDX-066",
+            ),
+            (
+                "permission-read",
+                "credentials-permissions-system-guard-destruction",
+                READ_EFFECTS,
+                "workspace",
+                ("operation-and-targets", "workspace-identity", "executable-identity"),
+                "review",
+                "context-required",
+                "CDX-066",
+            ),
+            (
+                "system-status",
+                "credentials-permissions-system-guard-destruction",
+                ("system-or-privilege-operation", "remote-state-read"),
+                "system",
+                ("operation-and-targets", "executable-identity"),
+                "review",
+                "context-required",
+                "CDX-066",
+            ),
+            (
+                "guard-uninstall-help",
+                "credentials-permissions-system-guard-destruction",
+                ("guard-control-operation",),
+                "guard",
+                ("operation-and-targets", "executable-identity"),
+                "review",
+                "decidable",
+                "CDX-066",
+            ),
+        )
+    }
+}
